update: 7/11/17 16:00

+++ app/src/main/java/com/example/shmuel/myapplication/cars/CarRecyclerViewAdapter.java
@@ -37,12 +37,12 @@ import java.util.ArrayList;
 
 public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerViewAdapter.ViewHolder> {
     BackEndFunc backEndFunc= FactoryMethod.getBackEndFunc(DataSourceType.DATA_LIST);
-    CarRecyclerViewAdapter.ViewHolder mHolder;
     private ArrayList<Car> objects;
     private Context mContext;
     int carId;
     public ActionMode actionMode;
     private int selectedPosition=-1;
+
     public CarRecyclerViewAdapter(ArrayList<Car> objects, Context context) {
         this.objects=objects;
         this.mContext=context;
@@ -58,27 +58,61 @@ public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerView
 
     @Override
     public void onBindViewHolder(final CarRecyclerViewAdapter.ViewHolder holder, final int position) {
-        this.mHolder=holder;
+        Car car = objects.get(position);
         if(selectedPosition==position){
             if(((MainActivity)mContext).is_in_action_mode==true){
-                holder.itemView.setBackgroundColor(Color.parseColor("#a3a3a3"));}
+                holder.itemView.setBackgroundColor(Color.parseColor("#a3a3a3"));
+                if(!car.isInUse())
+                {
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                        holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_not_in_use, mContext.getTheme()));
+                    } else {
+                        holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_not_in_use));
+                    }
+                }
+                else
+                {
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                        holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_in_use, mContext.getTheme()));
+                    } else {
+                        holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_in_use));
+                    }
+                }
+            }
         }
         else
+        {
+            holder.itemView.setBackgroundColor(Color.parseColor("#ffffff"));
+            if(!car.isInUse())
             {
-                holder.itemView.setBackgroundColor(Color.parseColor("#ffffff"));
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                    holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_not_in_use, mContext.getTheme()));
+                } else {
+                    holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_not_in_use));
+                }
             }
+            else
+            {
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                    holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_in_use, mContext.getTheme()));
+                } else {
+                    holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_in_use));
+                }
+            }
+        }
 
-            holder.itemView.setOnLongClickListener(new View.OnLongClickListener() {
+        holder.itemView.setOnLongClickListener(new View.OnLongClickListener() {
             @Override
             public boolean onLongClick(View v) {
                 MyActionModeCallbackCar callback=new MyActionModeCallbackCar();
                 actionMode=((Activity)mContext).startActionMode(callback);
                 actionMode.setTitle("delete car");
-                Toast.makeText(mContext,
-                        "long click", Toast.LENGTH_SHORT).show();
+
                 selectedPosition=position;
                 ((MainActivity)mContext).is_in_action_mode=true;
                 notifyDataSetChanged();
+                Toast.makeText(mContext,
+                        "long click", Toast.LENGTH_SHORT).show();
                 return true;
             }
         });
@@ -93,28 +127,17 @@ public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerView
             }
         });
 
-        Car car = objects.get(position);
+
         Address carAddress=backEndFunc.getBranch(car.getBranchNum()).getAddress();
         CarModel carModel=backEndFunc.getCarModel(car.getCarModel());
-
         int defaultImage = mContext.getResources().getIdentifier(car.getImgURL(),null,mContext.getPackageName());
         Drawable drawable= ContextCompat.getDrawable(mContext, defaultImage);
         holder.imageView.setImageDrawable(drawable);
         holder.branch.setText("Branch: "+carAddress.getStreet()+", "+carAddress.getCity());
-        //holder.carNumber.setText("#"+String.valueOf(car.getCarNum()));
-
         holder.carYear.setText(String.valueOf(car.getYear()));
-
-        //holder.city.setText(carAddress.getCity());
-        //holder.carModel.setText(carModel.getCarModelName());
         holder.companyName.setText(carModel.getCompanyName()+" "+carModel.getCarModelName());
-
-        //holder.street.setText(carAddress.getStreet());
-
-        //holder.milage.setText(String.valueOf(car.getMileage()));
         holder.dailyPrice.setText("USD "+String.valueOf(car.getOneDayCost()));
         holder.milePrice.setText("USD "+String.valueOf(car.getOneKilometerCost()));
-
         holder.ratingBar.setRating((float) car.getRating());
         holder.rating.setText(String.valueOf(car.getRating()));
         holder.numberOfRatings.setText("("+String.valueOf(car.getNumOfRatings())+")");
@@ -127,6 +150,14 @@ public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerView
                holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_not_in_use));
            }
        }
+       else
+       {
+           if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+               holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_in_use, mContext.getTheme()));
+           } else {
+               holder.inUse.setImageDrawable(mContext.getResources().getDrawable(R.drawable.ic_in_use));
+           }
+       }
 
 
     }
@@ -137,14 +168,8 @@ public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerView
     }
 
     static class ViewHolder extends RecyclerView.ViewHolder {
-        int carID;
         TextView companyName;
-        //TextView carModel;
-        //TextView carNumber;
         TextView carYear;
-        TextView city;
-        TextView street;
-        //TextView milage;
         TextView dailyPrice;
         TextView milePrice;
         ImageView imageView;
@@ -156,24 +181,15 @@ public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerView
 
         public ViewHolder(View itemView) {
             super(itemView);
-
             imageView=(ImageView)itemView.findViewById(R.id.carModelCardImage);
-
             companyName=(TextView)itemView.findViewById(R.id.carModelCardCompany);
-            //carModel=(TextView)itemView.findViewById(R.id.carModelCardModelName);
-            //carNumber=(TextView)itemView.findViewById(R.id.carCardId);
             branch=(TextView)itemView.findViewById(R.id.carCardBranch);
             carYear=(TextView)itemView.findViewById(R.id.carCardYear);
-            //city=(TextView)itemView.findViewById(R.id.carCardCity);
-           // street=(TextView)itemView.findViewById(R.id.carCardStreet);
-            //milage=(TextView)itemView.findViewById(R.id.carCardMilage);
             dailyPrice=(TextView)itemView.findViewById(R.id.carCardDayPrice);
             milePrice=(TextView)itemView.findViewById(R.id.carCardMilePrice);
-
             ratingBar=(RatingBar)itemView.findViewById(R.id.carCardratingBar);
             rating=(TextView)itemView.findViewById(R.id.carCardRating);
             numberOfRatings=(TextView)itemView.findViewById(R.id.carCardNumberOfRatings);
-
             inUse=(ImageButton)itemView.findViewById(R.id.carCardInUdeButton);
         }
     }
@@ -194,19 +210,24 @@ public class CarRecyclerViewAdapter extends RecyclerView.Adapter<CarRecyclerView
         public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
             switch (item.getItemId())
             {
-                case R.id.carDelete:{
+                case R.id.delete_item:{
+                    if(selectedPosition>-1 && objects.get(selectedPosition).isInUse()){
                     Toast.makeText(mContext,
-                            "it works!!!", Toast.LENGTH_SHORT).show();
-                    backEndFunc.deleteCar(objects.get(selectedPosition).getCarNum());
-                    notifyDataSetChanged();
+                            "cannot delete car, car is in use!!!", Toast.LENGTH_SHORT).show();
+                        return false;
+                    }
+                    else
+                    {
+                        backEndFunc.deleteCar(objects.get(selectedPosition).getCarNum());
+                        notifyDataSetChanged();
 
-                    selectedPosition=-1;
-                    notifyItemChanged(selectedPosition);
-                    actionMode.finish();
-                }
-                case android.R.id.closeButton:
-                {
-                    selectedPosition=-1;
+                        selectedPosition=-1;
+                        notifyItemChanged(selectedPosition);
+                        actionMode.finish();
+                        Toast.makeText(mContext,
+                                "car deleted", Toast.LENGTH_SHORT).show();
+                    }
+
                 }
             }
             return true;
